requirements.txt
asgiref==3.7.2
Django==4.2.6
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.8
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.3
